# openapi.yaml
openapi: 3.1.0
info:
  title: SchemaSprint BFF API
  version: 0.1.0
  description: Versioned same-origin contract for the private Web/PWA MVP. Provider traces and secrets are never returned.
servers:
  - url: /api
security:
  - sessionCookie: []
...
paths:
...
  /v1/problems/today:
    get:
      tags: [Problems]
      operationId: getTodayProblem
      parameters: [{$ref: '#/components/parameters/Locale'}]
      responses:
        '200': {description: Globally published problem for current JST practice date., content: {application/json: {schema: {$ref: '#/components/schemas/ProblemDetail'}}}}
        default: {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    sessionCookie: {type: apiKey, in: cookie, name: __Host-schemasprint_session}
  parameters:
...
    Locale: {name: locale, in: query, schema: {$ref: '#/components/schemas/Locale'}}
...
  responses:
    Error:
      description: Structured safe error.
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ProblemError'}}}
...
  schemas:
    Ulid: {type: string, pattern: '^[0-7][0-9A-HJKMNP-TV-Z]{25}$'}
    Locale: {type: string, enum: [en, ja, zh-CN, ko, es, pt-BR]}
    Difficulty: {type: string, enum: [easy, mid, hard]}
    InputFormat: {type: string, enum: [gui, mermaid, ddl, dbml, mixed]}
...
    ProblemDetail:
      type: object
      additionalProperties: false
      required: [id, version, title, difficulty, genre, inputFormat, tags, publicationDate, passed, submitted, translationStatus, problemVersionId, statement, officialAnswer, explanation, sourceLanguage, requestedLocale, rubricSummary]
      properties:
        id: {$ref: '#/components/schemas/Ulid'}
        version: {type: integer, minimum: 1}
        title: {type: string}
        difficulty: {$ref: '#/components/schemas/Difficulty'}
        genre: {type: string}
        inputFormat: {$ref: '#/components/schemas/InputFormat'}
        tags: {type: array, items: {type: string}}
        publicationDate: {type: string, format: date}
        passed: {type: boolean}
        submitted: {type: boolean}
        translationStatus: {type: string, enum: [source, translated, pending, unavailable]}
        problemVersionId: {$ref: '#/components/schemas/Ulid'}
        statement: {type: string}
        officialAnswer: {$ref: '#/components/schemas/CanonicalSchema'}
        explanation: {type: string}
        sourceLanguage: {type: string, const: en}
        requestedLocale: {$ref: '#/components/schemas/Locale'}
        rubricSummary:
          type: array
          maxItems: 2000
          items:
            type: object
            additionalProperties: false
            required: [id, description, weight, critical, implicit]
            properties:
              id: {$ref: '#/components/schemas/Ulid'}
              description: {type: string}
              weight: {type: integer, minimum: 1, maximum: 1000000}
              critical: {type: boolean}
              implicit: {type: boolean}
...
    CanonicalSchema:
      type: object
      additionalProperties: false
      required: [schemaVersion, tables, relationships]
      properties:
        schemaVersion: {type: integer, const: 1}
        tables:
          type: array
          maxItems: 200
          items: {$ref: '#/components/schemas/SchemaTable'}
        relationships:
          type: array
          maxItems: 2000
          items: {$ref: '#/components/schemas/SchemaRelationship'}
        enums: {type: array, maxItems: 200, items: {type: object}}
        views: {type: array, maxItems: 200, items: {type: object}}
        triggers: {type: array, maxItems: 200, items: {type: object}}
        policies: {type: array, maxItems: 500, items: {type: object}}
        partitions: {type: array, maxItems: 200, items: {type: object}}
    SchemaTable:
      type: object
      additionalProperties: false
      required: [id, name, columns, indexes, checks, position]
      properties:
        id: {type: string, minLength: 1, maxLength: 128}
        name: {type: string, minLength: 1, maxLength: 128}
        columns:
          type: array
          maxItems: 200
          items:
            type: object
            additionalProperties: false
            required: [id, name, dataType, nullable]
            properties:
              id: {type: string, minLength: 1, maxLength: 128}
              name: {type: string, minLength: 1, maxLength: 128}
              dataType: {type: string, minLength: 1, maxLength: 256}
              nullable: {type: boolean}
              primaryKey: {type: boolean, default: false}
              defaultExpression: {type: [string, 'null'], maxLength: 2000}
              generatedExpression: {type: [string, 'null'], maxLength: 2000}
        indexes: {type: array, maxItems: 200, items: {type: object}}
        checks: {type: array, maxItems: 200, items: {type: object}}
        position:
          type: object
          required: [x, y]
          properties: {x: {type: number}, y: {type: number}}
    SchemaRelationship:
      type: object
      additionalProperties: false
      required: [id, fromTableId, fromColumnIds, toTableId, toColumnIds, fromCardinality, toCardinality]
      properties:
        id: {type: string, minLength: 1, maxLength: 128}
        fromTableId: {type: string}
        fromColumnIds: {type: array, minItems: 1, maxItems: 32, items: {type: string}}
        toTableId: {type: string}
        toColumnIds: {type: array, minItems: 1, maxItems: 32, items: {type: string}}
        fromCardinality: {type: string, enum: ['0..1', '1', '0..*', '1..*']}
        toCardinality: {type: string, enum: ['0..1', '1', '0..*', '1..*']}
        onDelete: {type: string, enum: [no_action, restrict, cascade, set_null, set_default]}
...
    ProblemError:
      type: object
      additionalProperties: false
      required: [type, title, status, code, requestId]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        code: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}
        detail: {type: string, maxLength: 1000}
        requestId: {type: string}
        retryable: {type: boolean}
        retryAfterSeconds: {type: integer, minimum: 1, maximum: 86400}